refactor: split warnings.py into modular components
Extract warnings.py (1720 lines) into:
- lib/vtec_parser.py (113 lines) - VTEC/polygon parsing, zero Discord deps
- cogs/warning_format.py (530 lines) - Styling, formatting, URL generation
- cogs/warning_ui.py (560 lines) - Discord UI views (EnvironmentalView, TornadoPhotoView, TornadoDashboardView)
- cogs/warnings.py (360 lines) - WarningsCog class with polling logic

Reduces cognitive load per file, improves testability, enables reuse of
VTEC parsing in non-Discord contexts.
