+++ cogs/warnings.py
@@ -17,12 +17,11 @@ from __future__ import annotations
 import asyncio
 import json as _json
 import logging
-import os
 import re
 import time
-from datetime import datetime, timedelta, timezone
+from datetime import datetime, timezone
 from io import BytesIO
-from typing import List, Optional, Tuple
+from typing import Optional, Tuple
 
 import discord
 from discord import app_commands
@@ -30,6 +29,7 @@ from discord.ext import commands, tasks
 
 from config import NWS_ALERTS_WARNINGS_URL, WARNINGS_CHANNEL_ID
 from lib.vad_plotter.radar_coords import get_nearest_radar
+from lib.vtec_parser import parse_vtec, parse_warning_polygon, get_polygon_centroid
 from utils.backoff import TaskBackoff
 from utils.http import http_get_bytes, http_get_bytes_conditional
 from utils.state_store import (
@@ -39,1013 +39,23 @@ from utils.state_store import (
     get_recent_significant_events,
     prune_posted_warnings,
 )
+from cogs.warning_format import (
+    get_warning_style,
+    iem_autoplot_url,
+    build_concise_warning_text,
+    _vtec_url,
+    _vtec_unix_ts,
+    _area_with_state,
+    _extract_narrative,
+    _download_warning_image,
+)
+from cogs.warning_ui import (
+    EnvironmentalView,
+    TornadoPhotoView,
+    TornadoDashboardView,
+)
 
 logger = logging.getLogger("spc_bot")
-
-
-# ── VTEC parsing ─────────────────────────────────────────────────────────────
-
-# VTEC string format (NWS Directive 10-1703):
-#
-#   /O.NEW.KOUN.TO.W.0042.260427T2018Z-260427T2100Z/
-#    │ │   │    │  │ │    │
-#    │ │   │    │  │ │    issuance / expiration timestamps
-#    │ │   │    │  │ event tracking number (ETN, 4-digit, stable across the
-#    │ │   │    │  │ warning's full lifecycle — our dedup key)
-#    │ │   │    │  significance (W=warning, A=watch, Y=advisory, S=statement)
-#    │ │   │    phenomenon (TO=tornado, SV=svr tstm, FF=flash flood, etc.)
-#    │ │   issuing office (4-letter ICAO, e.g. KOUN = Norman)
-#    │ action (NEW, CON, EXP, CAN, UPG, EXA, EXT)
-#    fixed: O = operational
-_VTEC_RE = re.compile(
-    r"/O\.(NEW|CON|EXP|CAN|UPG|EXA|EXT|ROU)\."  # action
-    r"([A-Z]{4})\."                              # office
-    r"([A-Z]{2})\."                              # phenomenon
-    r"([A-Z])\."                                 # significance
-    r"(\d{4})\."                                 # ETN
-    r"(\d{6}T\d{4}Z)-(\d{6}T\d{4}Z)/"            # start/end
-)
-
-
-def parse_vtec(text: str) -> Optional[dict]:
-    """Parse the first VTEC string in ``text`` and return its components.
-
-    Returns a dict with ``action``, ``office``, ``phenom``, ``sig``,
-    ``etn``, plus the dedup key ``vtec_id`` (``OFFICE.PH.S.ETN``).
-    Returns ``None`` if no VTEC is present.
-    """
-    if not text:
-        return None
-    m = _VTEC_RE.search(text)
-    if not m:
-        return None
-    action, office, phenom, sig, etn, start, end = m.groups()
-    
-    # Normalize office to 4-letter ICAO (common for US WFOs in VTEC)
-    if len(office) == 3 and office[0] in "ABCDEFGHIJKLMNOPQRSTUVWXYZ":
-        office = "K" + office
-
-    return {
-        "action": action,
-        "office": office,
-        "phenom": phenom,
-        "sig": sig,
-        "etn": etn,
-        "start": start,
-        "end": end,
-        "vtec_id": f"{office}.{phenom}.{sig}.{etn}",
-    }
-
-
-# ── Polygon parsing (LAT...LON block) ────────────────────────────────────────
-
-_LATLON_RE = re.compile(
-    r"LAT\.\.\.LON\s+([\d\s]+?)(?=\n\s*[A-Z\$]|$)",
-    re.IGNORECASE | re.DOTALL,
-)
-
-
-def parse_warning_polygon(
-    text: str,
-) -> Optional[List[Tuple[float, float]]]:
-    """Parse the ``LAT...LON`` polygon block from a VTEC product.
-
-    Format: pairs of integer values, lat then lon, in degrees * 100,
-    space- or newline-delimited. Longitudes are reported as positive
-    integers; for the US they convert to negative decimal degrees.
-
-    Returns a list of (lat, lon) decimal-degree pairs, or ``None`` if
-    the block is missing or unparseable. Used by PR B's iembot
-    fallback to derive a polygon centroid when NWS API hasn't picked
-    up the alert yet.
-    """
-    if not text:
-        return None
-    m = _LATLON_RE.search(text)
-    if not m:
-        return None
-    nums = m.group(1).split()
-    coords: List[Tuple[float, float]] = []
-    for i in range(0, len(nums) - 1, 2):
-        try:
-            lat = int(nums[i]) / 100.0
-            lon = -(int(nums[i + 1]) / 100.0)
-        except ValueError:
-            continue
-        if not (15.0 <= lat <= 75.0 and -170.0 <= lon <= -60.0):
-            # Sanity-clip — NWS warnings only fire over US territory.
-            continue
-        coords.append((lat, lon))
-    return coords or None
-
-
-def get_polygon_centroid(
-    coords: List[Tuple[float, float]]
-) -> Optional[Tuple[float, float]]:
-    """Calculate the simple arithmetic centroid of a list of (lat, lon) pairs."""
-    if not coords:
-        return None
-    lat_sum = sum(c[0] for c in coords)
-    lon_sum = sum(c[1] for c in coords)
-    n = len(coords)
-    return (lat_sum / n, lon_sum / n)
-
-
-class EnvironmentalView(discord.ui.View):
-    def __init__(self, event_id: str):
-        super().__init__(timeout=None) # Persistent
-        self.event_id = event_id
-
-    @discord.ui.button(label="View Environmental Evolution", style=discord.ButtonStyle.secondary, emoji="📊", custom_id="view_env_evo")
-    async def view_env(self, interaction: discord.Interaction, button: discord.ui.Button):
-        await interaction.response.defer(ephemeral=True)
-        
-        from utils.events_db import get_events_db
-        db = await get_events_db()
-        async with db.execute(
-            "SELECT gif_path, srh_0_1, location FROM significant_events WHERE event_id = ?",
-            (self.event_id,)
-        ) as cur:
-            row = await cur.fetchone()
-
-        if not row or not row["gif_path"]:
-            # Check if there is an active mission
-            recorder = interaction.client.get_cog("RecorderCog")
-            is_active = False
-            if recorder:
-                # We'd need a way to check active missions by event_id
-                # For now, a generic message
-                await interaction.followup.send(
-                    "Environmental data is still being recorded or was not captured for this event. "
-                    "Try again in 90 minutes.", 
-                    ephemeral=True
-                )
-                return
-
-        # Post the GIF
-        gif_path = row["gif_path"]
-        if not os.path.exists(gif_path):
-             await interaction.followup.send("Archive file no longer exists on the server.", ephemeral=True)
-             return
-             
-        file = discord.File(gif_path, filename="evolution.gif")
-        embed = discord.Embed(
-            title=f"🌪️ Environmental Evolution - {row['location']}",
-            description=f"**Peak 0-1km SRH**: {row['srh_0_1']:.0f} m2/s2",
-            color=discord.Color.blue()
-        )
-        embed.set_image(url="attachment://evolution.gif")
-        await interaction.followup.send(embed=embed, file=file, ephemeral=True)
-
-
-# ── Cog ──────────────────────────────────────────────────────────────────────
-
-# (emoji, color) for each event type.
-_WARNING_STYLE = {
-    "Tornado Warning":             ("🌪️", discord.Color.red()),
-    "Severe Thunderstorm Warning": ("⛈️", discord.Color.gold()),
-    "Flash Flood Warning":         ("🌊", discord.Color.dark_blue()),
-    "Severe Weather Statement":    ("⛈️", discord.Color.gold()),
-    "Flash Flood Statement":       ("🌊", discord.Color.dark_blue()),
-    "Special Weather Statement":   ("☁️", discord.Color.blue()),
-}
-
-def get_warning_style(event: str, text: str, params: dict = None) -> Tuple[str, str, discord.Color, Optional[str]]:
-    """Determine (emoji, display_event_name, color, footer_id) based on event type and severity tags."""
-    emoji, color = _WARNING_STYLE.get(event, ("⚠️", discord.Color.orange()))
-    display_event = event
-    footer_id = None
-    
-    # Text-based detection (works for both iembot and NWS API paths)
-    text_upper = (text or "").upper()
-    
-    if event == "Tornado Warning":
-        if "TORNADO EMERGENCY" in text_upper:
-            return "🚨🚨", "Tornado Emergency", discord.Color.from_rgb(139, 0, 0), "EMERG"
-        if "PARTICULARLY DANGEROUS SITUATION" in text_upper:
-            return "⚠️", "Tornado Warning (PDS)", discord.Color.red(), "PDS"
-            
-    if event == "Severe Thunderstorm Warning":
-        if "THUNDERSTORM DAMAGE THREAT...DESTRUCTIVE" in text_upper:
-             return "🚨", "DESTRUCTIVE Severe Tstorm Warning", discord.Color.purple(), "EWX"
-        if "THUNDERSTORM DAMAGE THREAT...CONSIDERABLE" in text_upper:
-             return "⚠️", "CONSIDERABLE Severe Tstorm Warning", discord.Color.gold(), "EWX"
-
-    if event == "Flash Flood Warning":
-        if "FLASH FLOOD EMERGENCY" in text_upper:
-            return "🚨🚨", "Flash Flood Emergency", discord.Color.from_rgb(139, 0, 0), "EMERG"
-
-    # Param-based detection (NWS API specific)
-    if params:
-        t_threat = params.get("tornadoDamageThreat") or []
-        if "CATASTROPHIC" in t_threat:
-            return "🚨🚨", "Tornado Emergency", discord.Color.from_rgb(139, 0, 0), "EMERG"
-        if "CONSIDERABLE" in t_threat:
-            # Note: CONSIDERABLE tag for TOR usually means PDS
-            return "⚠️", "Tornado Warning (PDS)", discord.Color.red(), "PDS"
-
-        s_threat = params.get("thunderstormDamageThreat") or []
-        if "DESTRUCTIVE" in s_threat:
-             return "🚨", "DESTRUCTIVE Severe Tstorm Warning", discord.Color.purple(), "EWX"
-        if "CONSIDERABLE" in s_threat:
-             return "⚠️", "CONSIDERABLE Severe Tstorm Warning", discord.Color.gold(), "EWX"
-             
-        f_threat = params.get("flashFloodDamageThreat") or []
-        if "CATASTROPHIC" in f_threat:
-            return "🚨🚨", "Flash Flood Emergency", discord.Color.from_rgb(139, 0, 0), "EMERG"
-
-    return emoji, display_event, color, footer_id
-
-# Hard-cap on the description block we render inside a code-block —
-# Discord embed descriptions cap at 4096 chars total, fences add 8.
-_DESCRIPTION_LIMIT = 4000
-
-
-def iem_autoplot_url(vtec: dict) -> str:
-    """Return the IEM Autoplot URL (#208 for VTEC, #217 for SPS)."""
-    office = vtec["office"]
-    phenom = vtec["phenom"]
-    sig = vtec["sig"]
-    etn = vtec["etn"]
-    year = datetime.now(timezone.utc).year
-    if vtec.get("start"):
-        try:
-             year = 2000 + int(vtec["start"][:2])
-        except (ValueError, IndexError):
-             pass
-
-    # IEM expectations:
-    # 1. 3-letter SID for the WFO (e.g. KOUN -> OUN)
-    if office.startswith("K") and len(office) == 4:
-        office = office[1:]
-
-    # SPS (Special Weather Statements) use Autoplot 217 which requires the PID
-    if phenom == "SPS" and "-" in vtec["vtec_id"]:
-        return (
-            f"https://mesonet.agron.iastate.edu/plotting/auto/plot/217/"
-            f"pid:{vtec['vtec_id']}::segnum:0.png"
-        )
-
-    # Standard VTEC events use Autoplot 208
-    return (
-        f"https://mesonet.agron.iastate.edu/plotting/auto/plot/208/"
-        f"wfo:{office}::year:{year}::phenomena:{phenom}::significance:{sig}::"
-        f"etn:{etn.lstrip('0') or '0'}.png"
-    )
-
-
-async def _download_warning_image(image_url: str, filename: str) -> discord.File | None:
-    """Fetch an IEM Autoplot image with up to 8 attempts (~60s window).
-
-    Returns a ready-to-send discord.File, or None if all attempts fail.
-    Retries on 404 (IEM map not yet generated), 400 (bad request/pending),
-    or network errors with exponential backoff.
-    """
-    for attempt in range(8):
-        try:
-            content, status = await http_get_bytes(image_url, retries=1, timeout=15)
-            if content and status == 200:
-                return discord.File(BytesIO(content), filename=filename)
-
-            # Map might be pending (404/400). Use exponential backoff: 2s, 4s, 8s, 10s...
-            if attempt < 7:
-                delay = min(2 ** (attempt + 1), 10)
-                await asyncio.sleep(delay)
-                continue
-
-            logger.warning(
-                f"[WARN] Failed to download IEM image after 8 attempts: {image_url} (status={status})"
-            )
-        except Exception as e:
-            if attempt < 7:
-                delay = min(2 ** (attempt + 1), 10)
-                await asyncio.sleep(delay)
-                continue
-            logger.warning(f"[WARN] Error downloading IEM image after 8 attempts: {e}")
-        break
-    return None
-
-def _vtec_url(vtec: dict) -> str:
-    """Build an IEM VTEC event page URL from a parsed vtec dict."""
-    start = vtec.get("start", "")
-    if start and len(start) >= 11:
-        # '260429T0228Z' → '2026-04-29T02:28Z'
-        try:
-            year = 2000 + int(start[:2])
-            iso = f"{year}-{start[2:4]}-{start[4:6]}T{start[7:9]}:{start[9:11]}Z"
-        except (ValueError, IndexError):
-            now = datetime.now(timezone.utc)
-            year = now.year
-            iso = now.strftime("%Y-%m-%dT%H:%MZ")
-    else:
-        now = datetime.now(timezone.utc)
-        year = now.year
-        iso = now.strftime("%Y-%m-%dT%H:%MZ")
-    action = vtec.get("action", "NEW")
-    office = vtec.get("office", "")
-    phenom = vtec.get("phenom", "")
-    sig = vtec.get("sig", "")
-    etn = int(vtec.get("etn", "0") or "0")
-    return (
-        f"https://mesonet.agron.iastate.edu/vtec/f/"
-        f"{year}-O-{action}-{office}-{phenom}-{sig}-{etn:04d}_{iso}"
-    )
-
-
-def _vtec_unix_ts(vtec: dict) -> int:
-    """Return the Unix timestamp for the VTEC start time, or now if unavailable."""
-    start = vtec.get("start", "")
-    if start and len(start) >= 11:
-        try:
-            year = 2000 + int(start[:2])
-            month = int(start[2:4])
-            day = int(start[4:6])
-            hour = int(start[7:9])
-            minute = int(start[9:11])
-            return int(datetime(year, month, day, hour, minute, tzinfo=timezone.utc).timestamp())
-        except (ValueError, IndexError):
-            pass
-    return int(time.time())
-
-
-def _area_with_state(area_desc: str, ugc_codes: List[str]) -> str:
-    """Append [STATE] abbreviations to the area string, grouping counties by state.
-
-    Uses the NWS API geocode.UGC list (e.g. ['MSC023', 'ARC001']) to determine
-    which counties belong to which state, then formats them as:
-        'Clarke, Jasper, Jones [MS]'                         (single state)
-        'Ashley, Chicot [AR] and Washington [MS]'            (two states)
-    County names come from area_desc (already comma/semicolon separated).
-    The UGC ordering matches the area_desc ordering in NWS API responses.
-    """
-    if not ugc_codes:
-        return area_desc
-
-    # Parse county names from areaDesc
-    counties = [c.strip() for c in re.split(r'[;,]\s*', area_desc) if c.strip()]
-    if not counties:
-        return area_desc
-
-    # Group UGC codes by state (first 2 chars), preserving order of first appearance
-    from collections import OrderedDict
-    state_counts: dict = OrderedDict()
-    for ugc in ugc_codes:
-        if len(ugc) >= 2:
-            state = ugc[:2].upper()
-            state_counts[state] = state_counts.get(state, 0) + 1
-
-    if not state_counts:
-        return area_desc
-
-    # Split county list by state group counts
-    parts = []
-    idx = 0
-    for state, count in state_counts.items():
-        group = counties[idx:idx + count]
-        if group:
-            parts.append(f"{', '.join(group)} [{state}]")
-        idx += count
-
-    # Any leftover counties (mismatch in UGC/areaDesc lengths) appended to last group
-    if idx < len(counties):
-        remainder = counties[idx:]
-        if parts:
-            parts[-1] = parts[-1] + f", {', '.join(remainder)}"
-        else:
-            return area_desc
-
-    if len(parts) == 1:
-        return parts[0]
-    elif len(parts) == 2:
-        return f"{parts[0]} and {parts[1]}"
-    else:
-        return ", ".join(parts[:-1]) + f" and {parts[-1]}"
-
-
-def build_concise_warning_text(
-    display_event: str,
-    vtec: dict,
-    raw_text: Optional[str] = None,
-    feature: Optional[dict] = None,
-    ugc_codes: Optional[List[str]] = None,
-    is_update: bool = False,
-    prev_area: str = "",
-) -> str:
-    """Build the warning description for Discord.
-
-    Format: {office} [{verb} {display_event}](vtec_url) [{tags}] for {area} [STATE] till HH:MMZ
-            {narrative}
-            [<t:unix_ts:R>]
-    """
-    office = vtec["office"]
-    if office.startswith("K") and len(office) == 4:
-        office = office[1:]
-
-    # 1. Action verb
-    action_map = {
-        "NEW": "issues",
-        "CON": "continues",
-        "CAN": "cancels",
-        "EXP": "expires",
-        "EXT": "extends time of",
-        "UPG": "upgrades",
-    }
-    action_verb = action_map.get(vtec["action"], "updates")
-    if is_update:
-        action_verb = "updates"
-
-    # 2. Tags (tornado, hail, wind, flash flood)
-    tags = []
-    text_to_search = raw_text or ""
-    params = {}
-    if feature:
-        props = feature.get("properties", {})
-        text_to_search += " " + (props.get("description") or "")
-        params = props.get("parameters", {})
-
-    # Tornado Warning tags: [tornado: RADAR INDICATED, hail: 1.25 IN]
-    if "Tornado" in display_event:
-        if params.get("tornadoDetection"):
-            tags.append(f"tornado: {params['tornadoDetection'][0]}")
-        if params.get("tornadoDamageThreat"):
-            tags.append(f"damage threat: {params['tornadoDamageThreat'][0]}")
-        if params.get("maxHailSize"):
-            tags.append(f"hail: {params['maxHailSize'][0]} IN")
-
-    # Severe Thunderstorm Warning tags: [wind: 60 MPH (RADAR INDICATED), hail: 1.25 IN (RADAR INDICATED)]
-    elif "Severe Thunderstorm" in display_event:
-        w_method = ""
-        if params.get("windDetection"):
-            w_method = f" ({params['windDetection'][0]})"
-        if params.get("maxWindGust"):
-            tags.append(f"wind: {params['maxWindGust'][0]}{w_method}")
-            
-        h_method = ""
-        if params.get("hailDetection"):
-            h_method = f" ({params['hailDetection'][0]})"
-        if params.get("maxHailSize"):
-            tags.append(f"hail: {params['maxHailSize'][0]} IN{h_method}")
-
-        if params.get("thunderstormDamageThreat"):
-            tags.append(f"damage threat: {params['thunderstormDamageThreat'][0]}")
-
-    # Flash Flood Warning tags: [flash flood: radar indicated] (all lowercase)
-    elif "Flash Flood" in display_event:
-        if params.get("flashFloodDetection"):
-            tags.append(f"flash flood: {params['flashFloodDetection'][0].lower()}")
-        if params.get("flashFloodDamageThreat"):
-            tags.append(f"flash flood damage threat: {params['flashFloodDamageThreat'][0].lower()}")
-
-    # Fallback to regex if no params (iembot path)
-    if not tags and text_to_search:
-        m_tor = re.search(r"TORNADO\.\.\.(.+?)(?:\n|$)", text_to_search, re.I)
-        if m_tor:
-            tags.append(f"tornado: {m_tor.group(1).strip().upper()}")
-        m_hail = re.search(r"HAIL\.\.\.(.+?)(?:\n|$)", text_to_search, re.I)
-        if m_hail:
-            tags.append(f"hail: {m_hail.group(1).strip().upper()}")
-        m_wind = re.search(r"WIND\.\.\.(.+?)(?:\n|$)", text_to_search, re.I)
-        if m_wind:
-            tags.append(f"wind: {m_wind.group(1).strip().upper()}")
-
-    tag_str = f" [{', '.join(tags)}]" if tags else ""
-
-    # 3. Area (with [STATE] grouping when UGC codes are available)
-    area = "affected area"
-    if feature:
-        area = feature.get("properties", {}).get("areaDesc", area)
-    elif raw_text:
-        m_area = re.search(r"(?:Warning for|Statement for|IMPACT)\s+(.+?)(?=\n\s*\*|\n\s*At\s+|$)", raw_text, re.I | re.DOTALL)
-        if m_area:
-            raw_list = m_area.group(1)
-            parts = re.split(r"\n|\.\.\.|\s+AND\s+", raw_list, flags=re.I)
-            counties = []
-            for p in parts:
-                c = p.strip().strip(".")
-                if not c or len(c) < 3:
-                    continue
-                if any(x in c.upper() for x in ["THROUGH", "UNTIL", "PORTIONS", "AM", "PM", "EDT", "CDT", "MDT", "PDT", "HST", "AKDT"]):
-                    continue
-                c = re.sub(r"^(?:Northeastern|Northwestern|Southeastern|Southwestern|Northern|Southern|Eastern|Western|Central)\s+", "", c, flags=re.I)
-                c = re.split(r"\s+in\s+", c, flags=re.I)[0]
-                c = re.sub(r"\s+Count[iy].*$", "", c, flags=re.I)
-                c = c.strip()
-                if c and c.upper() not in ["CENTRAL", "NORTH", "SOUTH", "EAST", "WEST"] and c not in counties:
-                    counties.append(c)
-            if counties:
-                area = ", ".join(counties)
-
-    if is_update and prev_area:
-        # Calculate cancels/continues
-        prev_parts = [c.strip() for c in re.split(r'[;,]\s*', prev_area) if c.strip()]
-        curr_parts = [c.strip() for c in re.split(r'[;,]\s*', area) if c.strip()]
-        
-        prev_set = set(prev_parts)
-        curr_set = set(curr_parts)
-        
-        cancelled = sorted([c for c in prev_parts if c in (prev_set - curr_set)])
-        continuing = sorted([c for c in curr_parts if c in curr_set])
-        
-        if cancelled:
-            area_formatted = f" (**cancels** {', '.join(cancelled)}, **continues** {', '.join(continuing)})"
-        else:
-            area_formatted = f" for {_area_with_state(area, ugc_codes or [])}"
-    else:
-        area_formatted = f" for {_area_with_state(area, ugc_codes or [])}"
-
-    # 4. Expiration time (VTEC end field: '260428T0530Z')
-    expires_str = ""
-    if vtec.get("end"):
-        try:
-            z_time = vtec["end"].split("T")[1]
-            expires_str = f" till {z_time[:2]}:{z_time[2:4]}Z"
-        except (IndexError, ValueError):
-            pass
-
-    # 5. Narrative bullet
-    narrative = ""
-    if text_to_search:
-        # Narrative extraction: capture the "At..." bullet or the primary impact statement.
-        # We stop at the next bullet point (*) or known footer blocks.
-        m_nat = re.search(r"(?:\*\s*)?At\s+(.+?)(?=\n\s*\*|\n\s*LAT\.\.\.LON|\n\s*PRECAUTIONARY|\n\s*TIME\.\.\.MOT\.\.\.LOC|$)", text_to_search, re.I | re.DOTALL)
-        
-        # Fallback for Special Weather Statements which often lead with "...A STRONG THUNDERSTORM..."
-        if not m_nat and vtec.get("phenom") == "SPS":
-            # Capture from the first significant impact line starting with dots
-            m_nat = re.search(r"(?:\n|^)\s*\.\.\.([^.].+?)(?=\n\s*\*|\n\s*LAT\.\.\.LON|$)", text_to_search, re.I | re.DOTALL)
-
-        if m_nat:
-            val = m_nat.group(1).strip()
-            
-            # Refined bolding: only bold high-signal weather keywords followed by dots.
-            # Avoids bolding structural words like "NEAR...", "LOCATED...", or "TIME..."
-            def _bold_repl(m):
-                word = m.group(1).upper()
-                # Only bold if the word (excluding trailing dots) is a high-signal keyword
-                base_word = re.sub(r"\.+$", "", word)
-                if base_word in ("TORNADO", "HAIL", "WIND", "GUST", "WATERSPOUT", "IMPACT", "SOURCE", "MAX", "DAMAGE", "THREAT"):
-                    return f"**{m.group(1)}**"
-                return m.group(1)
-
-            val = re.sub(r"([A-Z]{4,}\b\.{3,})", _bold_repl, val)
-            val = re.sub(r"\s+", " ", val).strip()
-            val = val.lstrip(".").strip()
-            
-            # Use "At" prefix for standard warnings, or just the text for SPS
-            if "At" in m_nat.group(0):
-                 narrative = f"\nAt {val}"
-            else:
-                 narrative = f"\n{val}"
-
-    # 6. Hyperlinked verb + relative timestamp
-    vtec_link = _vtec_url(vtec)
-    unix_ts = _vtec_unix_ts(vtec)
-    linked_verb = f"[{action_verb} {display_event}]({vtec_link})"
-
-    # Period after area block for updates
-    suffix = "." if is_update else ""
-
-    return f"{office} {linked_verb}{tag_str}{area_formatted}{expires_str}{suffix}{narrative}\n[<t:{unix_ts}:R>]"
-
-
-def _extract_narrative(raw: str) -> Optional[str]:
-    """Pull the human-readable narrative out of a raw VTEC product.
-
-    The narrative is the section after the bulletin headers and before
-    the boilerplate footer (LAT...LON, ATTN, $$). Used by the iembot
-    fast-path when we don't yet have NWS API's pre-formatted description.
-    """
-    if not raw:
-        return None
-
-    text = raw
-    # Drop the WMO header / AFOS header / VTEC line block at the top so
-    # we lead with the substantive narrative rather than transmission
-    # metadata. Heuristic: find the line that begins "BULLETIN -" or
-    # the first line starting with "The National Weather Service in".
-    nws_idx = re.search(
-        r"(?m)^(?:BULLETIN.*|The National Weather Service\b.*)$", text
-    )
-    if nws_idx:
-        text = text[nws_idx.start():]
-
-    # Trim known footers — order matters because LAT...LON usually
-    # precedes ATTN.
-    for footer in ("LAT...LON", "ATTN...WFO", "TIME...MOT...LOC", "$$"):
-        m = re.search(re.escape(footer), text, re.IGNORECASE)
-        if m:
-            text = text[: m.start()]
-    text = text.strip()
-    return text or None
-
-
-class TornadoPhotoView(discord.ui.View):
-    def __init__(self, urls: list, parent_view: discord.ui.View, location: str):
-        super().__init__(timeout=300)
-        self.urls = urls
-        self.parent_view = parent_view
-        self.location = location
-        self.page = 0
-        self.per_page = 4
-        self._update_buttons()
-
-    def _update_buttons(self):
-        max_pages = (len(self.urls) + self.per_page - 1) // self.per_page
-        self.prev_btn.disabled = self.page <= 0
-        self.next_btn.disabled = self.page >= max_pages - 1
-
-    def build_embeds(self) -> list[discord.Embed]:
-        start = self.page * self.per_page
-        end = start + self.per_page
-        page_urls = self.urls[start:end]
-        
-        embeds = []
-        for i, url in enumerate(page_urls):
-            embed = discord.Embed(
-                title=f"📸 Damage Photos: {self.location}" if i == 0 else None,
-                color=discord.Color.teal(),
-                timestamp=datetime.now(timezone.utc) if i == 0 else None
-            )
-            # Use proxy URL or original URL
-            embed.set_image(url=url)
-            
-            if i == 0:
-                max_pages = (len(self.urls) + self.per_page - 1) // self.per_page
-                embed.set_footer(text=f"Page {self.page + 1} of {max_pages} ({len(self.urls)} total photos)")
-            
-            embeds.append(embed)
-            
-        return embeds
-
-    @discord.ui.button(label="◀ Prev Page", style=discord.ButtonStyle.secondary)
-    async def prev_btn(self, interaction: discord.Interaction, button: discord.ui.Button):
-        self.page = max(0, self.page - 1)
-        self._update_buttons()
-        await interaction.response.edit_message(embeds=self.build_embeds(), view=self)
-
-    @discord.ui.button(label="Next Page ▶", style=discord.ButtonStyle.secondary)
-    async def next_btn(self, interaction: discord.Interaction, button: discord.ui.Button):
-        max_pages = (len(self.urls) + self.per_page - 1) // self.per_page
-        self.page = min(max_pages - 1, self.page + 1)
-        self._update_buttons()
-        await interaction.response.edit_message(embeds=self.build_embeds(), view=self)
-
-    @discord.ui.button(label="🔙 Back to Card", style=discord.ButtonStyle.primary)
-    async def back_btn(self, interaction: discord.Interaction, button: discord.ui.Button):
-        await interaction.response.edit_message(embed=self.parent_view.build_card_embed(), embeds=[], view=self.parent_view)
-
-
-class TornadoDashboardView(discord.ui.View):
-    def __init__(self, events: list, title: str, mode: str = "card"):
-        super().__init__(timeout=300)
-        self.events = events
-        self.title = title
-        self.mode = mode # "summary" or "card"
-        self.index = 0   # Index for card mode
-        
-        # Group by date (UTC) for summary mode
-        self.grouped = {}
-        for e in events:
-            dt = datetime.fromtimestamp(e['timestamp'], timezone.utc)
-            date_str = dt.strftime("%Y-%m-%d")
-            if date_str not in self.grouped:
-                self.grouped[date_str] = []
-            self.grouped[date_str].append(e)
-            
-        self.dates = sorted(list(self.grouped.keys()), reverse=True)
-        self._update_items()
-
-    def _update_items(self):
-        self.clear_items()
-        
-        if self.mode == "summary":
-            # Build select options for summary
-            options = [discord.SelectOption(label="Summary Dashboard", value="summary", default=True)]
-            for d in self.dates[:24]:
-                options.append(discord.SelectOption(label=f"Events for {d}", value=d))
-            
-            select = discord.ui.Select(options=options, custom_id="date_select")
-            select.callback = self.on_select
-            self.add_item(select)
-        else:
-            # Card mode navigation
-            self.add_item(self.first_btn)
-            self.add_item(self.prev_btn)
-            self.add_item(self.next_btn)
-            self.add_item(self.last_btn)
-            self.add_item(self.summary_btn)
-
-            # Add Photos button if event has location and coords for DAT search
-            e = self.events[self.index]
-            if e.get("location") and e.get("coords"):
-                self.add_item(self.photos_btn)
-            
-            self.first_btn.disabled = self.index <= 0
-            self.prev_btn.disabled = self.index <= 0
-            self.next_btn.disabled = self.index >= len(self.events) - 1
-            self.last_btn.disabled = self.index >= len(self.events) - 1
-
-        # Global Archive Button
-        if self.events:
-            min_ts = min(e['timestamp'] for e in self.events)
-            max_ts = max(e['timestamp'] for e in self.events)
-            min_dt = datetime.fromtimestamp(min_ts, timezone.utc).strftime("%Y-%m-%d")
-            max_dt = (datetime.fromtimestamp(max_ts, timezone.utc) + timedelta(days=1)).strftime("%Y-%m-%d")
-            
-            # Stable URL format using query parameters
-            url = f"https://tornadoarchive.com/explorer/?start={min_dt}&end={max_dt}&domain=north_america"
-            self.add_item(discord.ui.Button(label="Tornado Archive", url=url, style=discord.ButtonStyle.link))
-
-    async def _render_map_if_needed(self, e: dict) -> Tuple[Optional[str], Optional[discord.File]]:
-        """Helper to fetch geometry and render a local map if possible."""
-        guid = e.get("dat_guid")
-        if not guid:
-            return None, None
-            
-        png_path = os.path.join("cache", f"track_{guid}.png")
-        if os.path.exists(png_path):
-            return f"attachment://track_{guid}.png", discord.File(png_path, filename=f"track_{guid}.png")
-
-        # Not cached, try to render now
-        from utils.dat_api import fetch_dat_track_geometry  # noqa: PLC0415
-        from utils.map_utils import render_tornado_track  # noqa: PLC0415
-        
-        try:
-            paths = await fetch_dat_track_geometry(guid)
-            if paths:
-                render_tornado_track(paths, png_path)
-                if os.path.exists(png_path):
-                    return f"attachment://track_{guid}.png", discord.File(png_path, filename=f"track_{guid}.png")
-        except Exception as err:
-            logger.warning(f"[DASHBOARD] Failed to render map for {guid}: {err}")
-            
-        return None, None
-
-    async def on_select(self, interaction: discord.Interaction):
-        val = interaction.data["values"][0]
-        if val == "summary":
-            await interaction.response.edit_message(embed=self.build_summary_embed(), embeds=[], view=self)
-        else:
-            # Switch to card mode at the first event of that day
-            day_events = self.grouped[val]
-            first_event = sorted(day_events, key=lambda x: x["timestamp"], reverse=True)[0]
-            self.index = self.events.index(first_event)
-            self.mode = "card"
-            self._update_items()
-            
-            # Check for map
-            img_url, file = await self._render_map_if_needed(first_event)
-            embed = self.build_card_embed()
-            if img_url:
-                embed.set_image(url=img_url)
-                embed.set_footer(text=f"Local DAT Render | {embed.footer.text}")
-                
-            await interaction.response.edit_message(embed=embed, view=self, attachments=[file] if file else [])
-
-    @discord.ui.button(label="⏮️ First", style=discord.ButtonStyle.secondary)
-    async def first_btn(self, interaction: discord.Interaction, button: discord.ui.Button):
-        self.index = 0
-        self._update_items()
-        e = self.events[self.index]
-        img_url, file = await self._render_map_if_needed(e)
-        embed = self.build_card_embed()
-        if img_url:
-            embed.set_image(url=img_url)
-            embed.set_footer(text=f"Local DAT Render | {embed.footer.text}")
-        await interaction.response.edit_message(embed=embed, view=self, attachments=[file] if file else [])
-
-    @discord.ui.button(label="◀ Prev", style=discord.ButtonStyle.secondary)
-    async def prev_btn(self, interaction: discord.Interaction, button: discord.ui.Button):
-        self.index = max(0, self.index - 1)
-        self._update_items()
-        e = self.events[self.index]
-        img_url, file = await self._render_map_if_needed(e)
-        embed = self.build_card_embed()
-        if img_url:
-            embed.set_image(url=img_url)
-            embed.set_footer(text=f"Local DAT Render | {embed.footer.text}")
-        await interaction.response.edit_message(embed=embed, view=self, attachments=[file] if file else [])
-
-    @discord.ui.button(label="Next ▶", style=discord.ButtonStyle.secondary)
-    async def next_btn(self, interaction: discord.Interaction, button: discord.ui.Button):
-        self.index = min(len(self.events) - 1, self.index + 1)
-        self._update_items()
-        e = self.events[self.index]
-        img_url, file = await self._render_map_if_needed(e)
-        embed = self.build_card_embed()
-        if img_url:
-            embed.set_image(url=img_url)
-            embed.set_footer(text=f"Local DAT Render | {embed.footer.text}")
-        await interaction.response.edit_message(embed=embed, view=self, attachments=[file] if file else [])
-
-    @discord.ui.button(label="Last ⏭️", style=discord.ButtonStyle.secondary)
-    async def last_btn(self, interaction: discord.Interaction, button: discord.ui.Button):
-        self.index = len(self.events) - 1
-        self._update_items()
-        e = self.events[self.index]
-        img_url, file = await self._render_map_if_needed(e)
-        embed = self.build_card_embed()
-        if img_url:
-            embed.set_image(url=img_url)
-            embed.set_footer(text=f"Local DAT Render | {embed.footer.text}")
-        await interaction.response.edit_message(embed=embed, view=self, attachments=[file] if file else [])
-
-    @discord.ui.button(label="📋 Summary", style=discord.ButtonStyle.primary)
-    async def summary_btn(self, interaction: discord.Interaction, button: discord.ui.Button):
-        self.mode = "summary"
-        self._update_items()
-        await interaction.response.edit_message(embed=self.build_summary_embed(), view=self)
-
-    @discord.ui.button(label="📸 Photos", style=discord.ButtonStyle.success)
-    async def photos_btn(self, interaction: discord.Interaction, button: discord.ui.Button):
-        e = self.events[self.index]
-
-        # Need location and coords to search DAT
-        location = e.get("location")
-        coords = e.get("coords")
-        event_id = e.get("event_id")
-        if not location or not coords:
-            await interaction.response.send_message("No location data available for this event.", ephemeral=True)
-            return
-
-        await interaction.response.defer()
-        from utils.events_db import fetch_dat_photos, get_cached_dat_photos
-
-        # Check cache first (instant load)
-        photos = []
-        if event_id:
-            photos = get_cached_dat_photos(event_id)
-
-        # If not cached, fetch from DAT
-        if not photos:
-            magnitude = e.get("magnitude", "")
-            photo_urls = await fetch_dat_photos(
-                location=location,
-                magnitude=magnitude,
-                coords=coords,
-            )
-            if photo_urls:
-                # Convert URLs to file paths if possible, otherwise use URLs
-                photos = photo_urls
-        else:
-            logger.info(f"[WARNINGS] Using {len(photos)} cached photo(s) for {event_id}")
-
-        if not photos:
-            await interaction.followup.send("No damage photos found in the DAT for this event.", ephemeral=True)
-            return
-
-        photo_view = TornadoPhotoView(photos, self, location)
-        await interaction.edit_original_response(embeds=photo_view.build_embeds(), view=photo_view)
-
-    def _get_ef_emoji(self, mag: str) -> str:
-        mag = (mag or "").upper()
-        if "EF5" in mag: return "🟣"
-        if "EF4" in mag: return "🔴"
-        if "EF3" in mag: return "🟠"
-        if "EF2" in mag: return "🟡"
-        if "EF1" in mag: return "🟢"
-        if "EF0" in mag: return "🔵"
-        return "⚪"
-
-    def build_summary_embed(self) -> discord.Embed:
-        # Calculate Grand Totals
-        totals = {"EF5": 0, "EF4": 0, "EF3": 0, "EF2": 0, "EF1": 0, "EF0": 0, "EFU": 0}
-        for e in self.events:
-            mag = (e.get("magnitude") or "").upper()
-            matched = False
-            for scale in ["EF5", "EF4", "EF3", "EF2", "EF1", "EF0"]:
-                if scale in mag:
-                    totals[scale] += 1
-                    matched = True
-                    break
-            if not matched:
-                totals["EFU"] += 1
-
-        total_count = len(self.events)
-        
-        # Build total line: 🟣0 🔴0 🟠2 🟡5 🟢14 🔵21 ⚪6
-        t_parts = []
-        if totals["EF5"]: t_parts.append(f"🟣{totals['EF5']}")
-        if totals["EF4"]: t_parts.append(f"🔴{totals['EF4']}")
-        if totals["EF3"]: t_parts.append(f"🟠{totals['EF3']}")
-        if totals["EF2"]: t_parts.append(f"🟡{totals['EF2']}")
-        if totals["EF1"]: t_parts.append(f"🟢{totals['EF1']}")
-        if totals["EF0"]: t_parts.append(f"🔵{totals['EF0']}")
-        if totals["EFU"]: t_parts.append(f"⚪{totals['EFU']}")
-        
-        total_line = " ".join(t_parts) if t_parts else "None"
-        
-        lines = [
-            f"**Total: {total_line} ({total_count} tornadoes)**",
-            "───────────────────────────"
-        ]
-        
-        # Add daily breakdown
-        for date_str in self.dates[:20]: # Limit to 20 days for description length
-            day_events = self.grouped[date_str]
-            d_counts = {"EF5": 0, "EF4": 0, "EF3": 0, "EF2": 0, "EF1": 0, "EF0": 0, "EFU": 0}
-            for e in day_events:
-                mag = (e.get("magnitude") or "").upper()
-                matched = False
-                for scale in ["EF5", "EF4", "EF3", "EF2", "EF1", "EF0"]:
-                    if scale in mag:
-                        d_counts[scale] += 1
-                        matched = True
-                        break
-                if not matched:
-                    d_counts["EFU"] += 1
-
-            # Shorten date: 2026-05-01 -> May 01
-            try:
-                dt_obj = datetime.strptime(date_str, "%Y-%m-%d")
-                short_date = dt_obj.strftime("%b %d")
-            except ValueError:
-                short_date = date_str
-
-            # Build daily parts
-            d_parts = []
-            if d_counts["EF5"]: d_parts.append(f"🟣{d_counts['EF5']}")
-            if d_counts["EF4"]: d_parts.append(f"🔴{d_counts['EF4']}")
-            if d_counts["EF3"]: d_parts.append(f"🟠{d_counts['EF3']}")
-            if d_counts["EF2"]: d_parts.append(f"🟡{d_counts['EF2']}")
-            if d_counts["EF1"]: d_parts.append(f"🟢{d_counts['EF1']}")
-            if d_counts["EF0"]: d_parts.append(f"🔵{d_counts['EF0']}")
-            if d_counts["EFU"]: d_parts.append(f"⚪{d_counts['EFU']}")
-            
-            day_icons = " ".join(d_parts) if d_parts else "Confirmed"
-            # Use fixed-width date for alignment in monospace blocks
-            lines.append(f"`{short_date} ({len(day_events):>2})` .... {day_icons}")
-            
-        embed = discord.Embed(
-            title=f"{self.title}",
-            description="\n".join(lines),
-            color=discord.Color.red(),
-            timestamp=datetime.now(timezone.utc)
-        )
-        
-        embed.set_footer(
-            text=f"Showing last {min(20, len(self.dates))} active days. Use dropdown to pick a day."
-        )
-        return embed
-        
-    def build_card_embed(self) -> discord.Embed:
-        e = self.events[self.index]
-        dt = datetime.fromtimestamp(e['timestamp'], timezone.utc)
-        date_str = dt.strftime("%Y-%m-%d %H:%MZ")
-        rel_time = f"<t:{int(e['timestamp'])}:R>"
-        
-        mag = e.get("magnitude", "Confirmed")
-        emoji = self._get_ef_emoji(mag)
-        
-        embed = discord.Embed(
-            title=f"{emoji} Tornado: {e['location']}",
-            color=discord.Color.red(),
-            timestamp=datetime.now(timezone.utc)
-        )
-        
-        embed.add_field(name="Rating", value=mag, inline=True)
-        embed.add_field(name="Time", value=f"{date_str}\n({rel_time})", inline=True)
-        embed.add_field(name="Office", value=e['source'], inline=True)
-        
-        if e.get("lead_time") is not None:
-             embed.add_field(name="Lead Time", value=f"{e['lead_time']:.1f} min", inline=True)
-        
-        if e.get("vtec_id"):
-            parts = e["vtec_id"].split(".")
-            if len(parts) == 4:
-                office, phenom, sig, etn = parts
-                url = _vtec_url({
-                    "action": "NEW",
-                    "office": office,
-                    "phenom": phenom,
-                    "sig": sig,
-                    "etn": etn,
-                    "start": dt.strftime("%y%m%dT%H%MZ"),
-                })
-                embed.add_field(name="VTEC ID", value=f"[{e['vtec_id']}]({url})", inline=True)
-
-        if e.get("dat_guid"):
-            dat_url = f"https://apps.dat.noaa.gov/stormdamage/damageviewer/?datglobalid={e['dat_guid']}"
-            embed.add_field(name="NWS DAT", value=f"[View Track]({dat_url})", inline=True)
-            
-            # Show track map in the card
-            # We use IEM Autoplot for the dashboard view as it handles historical caching better
-            event_date = dt.strftime("%Y-%m-%d")
-            img_url = (
-                f"https://mesonet.agron.iastate.edu/plotting/auto/plot/253/"
-                f"datglobalid:{e['dat_guid']}::dat:{event_date}::cmap:gist_rainbow::"
-                f"_r:t::dpi:100.png"
-            )
-            embed.set_image(url=img_url)
-
-        if e.get("raw_text"):
-             text = e["raw_text"]
-             if len(text) > 500:
-                  text = text[:497] + "..."
-             embed.add_field(name="Remarks", value=f"```\n{text}\n```", inline=False)
-            
-        embed.set_footer(text=f"Event {self.index + 1} of {len(self.events)} | {e['coords']}")
-        return embed
 
 
 class WarningsCog(commands.Cog):
@@ -1168,7 +178,7 @@ class WarningsCog(commands.Cog):
             # For updates, we don't overwrite the dict yet, but we mark the product.
             self.bot.state.posted_product_ids.add(product_id)
             if not is_update:
-                self.bot.state.posted_warnings[vtec_id] = {} # placeholder
+                self.bot.state.posted_warnings[vtec_id] = {}  # placeholder
 
             self.bot.state.active_warnings[vtec_id] = vtec
 
@@ -1237,30 +247,31 @@ class WarningsCog(commands.Cog):
                 # Keep product IDs pruned to today's set (roughly)
                 if len(self.bot.state.posted_product_ids) > 1000:
                     # Very crude pruning — ideally we'd use a TTL but this is in-memory
-                    self.bot.state.posted_product_ids.clear() 
+                    self.bot.state.posted_product_ids.clear()
                 await prune_posted_warnings()
             except Exception as e:
                 logger.warning(f"[WARN] Failed to persist {vtec_id}: {e}")
         finally:
             self._in_flight_vtecs.discard(vtec_id)
+
     async def _check_and_log_significant_event(self, event: str, raw_text: str, vtec: dict):
         """Parse warning text for confirmed tornadoes and log to DB."""
         text_upper = (raw_text or "").upper()
         vtec_id = vtec.get("vtec_id", "Unknown")
-        
+
         # 1. Confirmed Tornado Detection
         # Check for OBSERVED tag or CONFIRMED wording
         is_confirmed = False
         if "TORNADO...OBSERVED" in text_upper or "CONFIRMED TORNADO" in text_upper:
             is_confirmed = True
-        
+
         if event == "Tornado Warning" and is_confirmed:
             # Extract location (rough approximation from first line of narrative)
             location = "Unknown Area"
             m_area = re.search(r"(?:near|over)\s+(.+?)(?:,)", raw_text, re.I)
             if m_area:
                 location = m_area.group(1).strip()
-            
+
             # Extract coords
             coords = ""
             m_poly = re.search(r"LAT\.\.\.LON\s+(.+?)(?=\n|\$\$|$)", raw_text, re.DOTALL)
@@ -1270,7 +281,7 @@ class WarningsCog(commands.Cog):
             office = vtec.get("office", "NWS")
             from utils.state_store import find_matching_tornado
             match_id = await find_matching_tornado(office, time.time(), location, window_hours=1.0)
-            
+
             event_id = match_id if match_id else f"NWS:WARN:{vtec_id}"
 
             await add_significant_event(
@@ -1301,7 +312,7 @@ class WarningsCog(commands.Cog):
                 logger.warning(f"[WARN] Failed to trigger VAD recorder for {vtec_id}: {e}")
 
             return event_id
-        
+
         return None
 
     @app_commands.command(name="recenttornadoes", description="List confirmed tornadoes from recent warnings and reports")
@@ -1319,7 +330,7 @@ class WarningsCog(commands.Cog):
     ])
     async def recent_tornadoes(self, interaction: discord.Interaction, range: int = 24):
         await interaction.response.defer()
-        
+
         events = await get_recent_significant_events(event_type="Tornado", since_hours=range)
         if not events:
             await interaction.followup.send("No confirmed tornadoes logged in the requested time frame.")
@@ -1327,10 +338,10 @@ class WarningsCog(commands.Cog):
 
         # Sort by timestamp DESC just in case
         events.sort(key=lambda x: x["timestamp"], reverse=True)
-        
+
         view = TornadoDashboardView(events, f"🌪️ Confirmed Tornadoes (Last {range}h)")
         embed = view.build_summary_embed()
-        
+
         await interaction.followup.send(embed=embed, view=view)
 
     @app_commands.command(name="sigtor", description="List significant (EF2+) tornadoes from recent surveys")
@@ -1344,7 +355,7 @@ class WarningsCog(commands.Cog):
     ])
     async def sig_tor(self, interaction: discord.Interaction, range: int = 168):
         await interaction.response.defer()
-        
+
         events = await get_recent_significant_events(event_type="Tornado", since_hours=range)
         if not events:
             await interaction.followup.send("No confirmed tornadoes logged in the requested time frame.")
@@ -1360,7 +371,7 @@ class WarningsCog(commands.Cog):
                 is_sig = True
             elif "SIGNIFICANT" in mag or "PDS" in mag:
                 is_sig = True
-            
+
             if is_sig:
                 sig_events.append(e)
 
@@ -1370,10 +381,10 @@ class WarningsCog(commands.Cog):
 
         # Sort by timestamp DESC
         sig_events.sort(key=lambda x: x["timestamp"], reverse=True)
-        
+
         view = TornadoDashboardView(sig_events, f"🚨 Significant Tornadoes (Last {range}h)")
         embed = view.build_summary_embed()
-        
+
         await interaction.followup.send(embed=embed, view=view)
 
     # phenom+sig → human-readable event name, for cancellation posts
@@ -1419,7 +430,7 @@ class WarningsCog(commands.Cog):
         # without the text or params.
         event_base = self._PHENOM_EVENT.get((phenom, sig), f"{phenom}.{sig} Warning")
         _, display_event, _, footer_id = get_warning_style(event_base, "")
-        
+
         action_verb = "cancels" if reason == "Cancelled" else "expires"
         area_str = f" for {area}" if area else ""
 
@@ -1457,7 +468,7 @@ class WarningsCog(commands.Cog):
         )
         if files:
             embed.set_image(url=f"attachment://{files[0].filename}")
-        
+
         footer_text = f"VTEC {vtec_id}"
         if footer_id:
             footer_text += f" | {footer_id}"
@@ -1468,7 +479,6 @@ class WarningsCog(commands.Cog):
             self._cancelled_warnings.add(vtec_id)
             logger.info(f"[WARN] Posted cancellation for {vtec_id}")
         except Exception as e:
-
             logger.warning(f"[WARN] Failed to post cancellation for {vtec_id}: {e}")
 
     @tasks.loop(seconds=30)
@@ -1521,6 +531,7 @@ class WarningsCog(commands.Cog):
 
         current_vtec_data = {}
         current_vtec_ids = set()
+        from cogs.warning_format import _WARNING_STYLE
         for feature in alert_response.features:
             props = feature.properties
             event = props.event
@@ -1538,7 +549,7 @@ class WarningsCog(commands.Cog):
                         break
             if not vtec_dict:
                 continue
-            
+
             issuance_id = vtec_dict["vtec_id"]
             # Store the vtec dict so disappeared path can use it for graphics
             current_vtec_data[issuance_id] = vtec_dict
@@ -1556,13 +567,13 @@ class WarningsCog(commands.Cog):
                 # Still active, ensures it stays in the active set
                 if issuance_id not in self.bot.state.active_warnings:
                     self.bot.state.active_warnings[issuance_id] = vtec_dict
-                
+
                 # Check for area change (partial cancellation) on CON products
                 if vtec_dict["action"] == "CON":
                     stored_info = self.bot.state.posted_warnings[issuance_id]
                     prev_area = stored_info.get("area", "")
                     curr_area = props.areaDesc or ""
-                    
+
                     if prev_area and curr_area and prev_area != curr_area:
                         # Area changed - likely a partial cancellation
                         if issuance_id not in self._in_flight_vtecs:
@@ -1572,13 +583,13 @@ class WarningsCog(commands.Cog):
                                     await self._post_warning(feature, channel, vtec_dict, event, is_update=True)
                                 except discord.HTTPException as e:
                                     logger.exception(f"[WARN] Update send failed for {issuance_id}: {e}")
-                                
+
                                 # Update stored area so we don't spam updates for every poll
                                 self.bot.state.posted_warnings[issuance_id]["area"] = curr_area
                                 await add_posted_warning(
-                                    issuance_id, 
-                                    stored_info["message_id"], 
-                                    stored_info["channel_id"], 
+                                    issuance_id,
+                                    stored_info["message_id"],
+                                    stored_info["channel_id"],
                                     area=curr_area
                                 )
                             finally:
@@ -1589,7 +600,7 @@ class WarningsCog(commands.Cog):
             # Check in-flight set to avoid racing with concurrent NWWS triggers
             if issuance_id in self._in_flight_vtecs:
                 continue
-            
+
             # Final check of posted_warnings to be absolutely sure
             if issuance_id in self.bot.state.posted_warnings:
                 continue
@@ -1603,7 +614,7 @@ class WarningsCog(commands.Cog):
                     continue
 
                 # Claim the dedup key BEFORE any awaits
-                self.bot.state.posted_warnings[issuance_id] = {} # placeholder
+                self.bot.state.posted_warnings[issuance_id] = {}  # placeholder
 
                 try:
                     msg, area_desc = await self._post_warning(feature, channel, vtec_dict, event)
@@ -1661,7 +672,7 @@ class WarningsCog(commands.Cog):
 
         ugc_codes = (props.geocode.UGC or []) if props.geocode else []
         area_desc = _area_with_state(props.areaDesc or "", ugc_codes)
-        
+
         prev_area = ""
         if is_update:
             prev_area = self.bot.state.posted_warnings.get(vtec_id, {}).get("area", "")
